# product-forge/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Product Forge API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/validation": {
      "get": {
        "operationId": "validation_api_v1_validation_get",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validation",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/feature-pr": {
      "post": {
        "operationId": "feature_pr_api_v1_validation_feature_pr_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Pr",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/integration": {
      "post": {
        "operationId": "integration_api_v1_validation_integration_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Integration",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/merge-gate": {
      "get": {
        "operationId": "merge_gate_api_v1_validation_merge_gate_get",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "project",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Gate",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/merge-gate/queue": {
      "get": {
        "operationId": "merge_queue_api_v1_validation_merge_gate_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "project",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "",
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Queue",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/policy": {
      "get": {
        "operationId": "policy_api_v1_validation_policy_get",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Policy",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/profiles": {
      "get": {
        "operationId": "validation_profiles_api_v1_validation_profiles_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validation Profiles",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/run": {
      "post": {
        "operationId": "run_validation_api_v1_validation_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Validation",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/validation/runs": {
      "get": {
        "operationId": "validation_runs_api_v1_validation_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "project",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "default": "",
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "default": "",
              "title": "Profile",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-roles",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Roles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validation Runs",
        "tags": [
          "validation"
        ]
      }
    },
...
  }
}